+++ src/main/java/com/proyecto/app/restcontroller/UsuarioController.java
@@ -16,6 +16,8 @@ import javax.validation.Valid;
 import com.proyecto.app.dto.UsuarioDto;
 import com.proyecto.app.entity.Usuario;
 import com.proyecto.app.service.IUsuarioService;
+import com.proyecto.app.wrapper.OrganizacionWrapper;
+import com.proyecto.app.wrapper.UsuarioWrapper;
 
 @RequestMapping("api/usuarios")
 @RestController
@@ -29,40 +31,42 @@ public class UsuarioController {
 	@GetMapping("/todosUsuarios")
 	public ResponseEntity<Map<String, Object>> all(){
 		Map<String, Object> response = new HashMap<>();
-		List<Usuario> usuarios = new ArrayList<>();
-		response.put("usuarios", usuarios);
-		return new ResponseEntity<Map<String, Object>>(response, HttpStatus.CREATED);
+		List<Usuario> usuarios = usuarioService.getAll();
+		response.put("Usuarios: ", usuarios);
+		return new ResponseEntity<Map<String, Object>>(response, HttpStatus.OK);
 
 	}
 	
-	@GetMapping("/usuario/{id}")
-	public ResponseEntity<HashMap<String, Object>> usuario(@PathVariable(value = "id") Long id){
-		HashMap<String, Object> response = new HashMap<String, Object>();
+	@GetMapping("/usuarioById/{id}")
+	public ResponseEntity<Map<String, Object>> usuario(@PathVariable(name = "id") Long id){
+		Map<String, Object> response = new HashMap<>();
 		Usuario usuario = usuarioService.findById(id);
-		response.put("usuario: ", usuario);
-		return new ResponseEntity<HashMap<String, Object>>(response, HttpStatus.OK);
+		response.put("Usuario: ", usuario);
+		return new ResponseEntity<Map<String, Object>>(response, HttpStatus.OK);
 	}
 	
-	@GetMapping("/usuario/{lastname}")
-	public ResponseEntity<HashMap<String, Object>> buscarPorApellido(@PathVariable(name = "name") String lastname) throws NotFoundException{
-		HashMap<String, Object> response = new HashMap<String, Object>();
-
-		if(lastname.equals("Rey")) {
-			throw new NotFoundException();
+	@GetMapping("/usuarioByLastname/{lastname}")
+	public ResponseEntity<Map<String, Object>> findByLastname(@PathVariable(name = "lastname") String lastname) throws NotFoundException{
+		Map<String, Object> response = new HashMap<>();
+		List<Usuario> usuarios = usuarioService.findByLastname(lastname);
+		if(usuarios!=null) {
+			response.put("Usuario: ", usuarios);
+		}else {
+			response.put("Mensaje", "El usuario no existe!");
 		}
-		response.put("mensaje", "El usuario con apellido ".concat(lastname).concat(" existe"));
-		return new ResponseEntity<>(response, HttpStatus.OK);
+		return new ResponseEntity<Map<String,Object>>(response, HttpStatus.OK);
 	}
 	
-	@GetMapping("/usuario/{dni}")
-	public ResponseEntity<HashMap<String, Object>> buscarPorDni(@PathVariable(name = "name") String dni) throws NotFoundException{
-		HashMap<String, Object> response = new HashMap<String, Object>();
-
-		if(dni.equals("39190124")) {
-			throw new NotFoundException();
+	@GetMapping("/usuarioByDni/{dni}")
+	public ResponseEntity<Map<String, Object>> buscarPorDni(@PathVariable(name = "dni") Long dni) throws NotFoundException{
+		Map<String, Object> response = new HashMap<>();
+		UsuarioDto usuario = usuarioService.findByDni(dni);
+		if(usuario!=null) {
+			response.put("Usuario: ", usuario);
+		}else {
+			response.put("Mensaje", "El usuario no existe!");
 		}
-		response.put("mensaje", "El usuario con DNI: ".concat(dni).concat(" existe"));
-		return new ResponseEntity<>(response, HttpStatus.OK);
+		return new ResponseEntity<Map<String,Object>>(response, HttpStatus.OK);
 	}
 	
 	@PostMapping("/registro")
@@ -71,33 +75,36 @@ public class UsuarioController {
 		log.info("usuario: "+usuarioDto.toString());
 		Map<String, Object> response = new HashMap<>();
 		UsuarioDto newUsuario = usuarioService.save(usuarioDto);
-		response.put("usuario: ", newUsuario);
+		response.put("Usuario: ", newUsuario);
 		return new ResponseEntity<Map<String,Object>>(response, HttpStatus.OK);
 	}
 	
-	@PutMapping("/updateUser{id}")
-	public ResponseEntity<Map<String, Object>> update(@RequestBody UsuarioDto usuarioDto){
+	@PutMapping("/updateUser")
+	public ResponseEntity<Map<String, Object>> update(@RequestParam(value="clave",required = true) String clave,@RequestBody UsuarioDto usuarioDto){
 		log.info("usuario: "+usuarioDto.toString());
 		Map<String, Object> response = new HashMap<>();
-		UsuarioDto updateUser = usuarioService.save(usuarioDto);
-		
-		if(updateUser == null) {
+		Usuario usuario = usuarioService.findByClave(clave);
+		if(usuario!=null) {
+		UsuarioDto updateUser = usuarioService.update(usuarioDto);
+		response.put("Usuario: ", updateUser);
+		}else{
 			response.put("mensaje", "No se pudo actualizar la informacion del usuario.");
 		}
 		
-		response.put("usuario: ", updateUser);
 		return new ResponseEntity<Map<String,Object>>(response, HttpStatus.OK);
 	}
 	
-	@DeleteMapping("/delete/{dni}")
-	  public ResponseEntity<Map<String, Object>> delete(@PathVariable UsuarioDto usuarioDto) {
+	@DeleteMapping("/delete")
+	  public ResponseEntity<Map<String, Object>> delete(@RequestParam(value="clave",required = true) String clave) {
 		Map<String, Object> response = new HashMap<>();
+		Usuario usuario = usuarioService.findByClave(clave);
+		if(usuario!=null) {
+		UsuarioDto usuarioDto = UsuarioWrapper.entityToDto(usuario);
 		UsuarioDto updateUser = usuarioService.delete(usuarioDto);
-		if(updateUser == null) {
+		response.put("Usuario: ", updateUser);
+		}else {
 			response.put("mensaje", "No se pudo borrar la informacion del usuario.");
 		}
-		
-		response.put("usuario: ", updateUser);
 		
 		return new ResponseEntity<Map<String,Object>>(response, HttpStatus.OK);
 	  }
